refactored out createUrqlClient into...
a util folder to clean up _app
added withUrql to wrap our pages and at the moment ssr rendering /
page.  turned off on login and register.

made a util function to check if window is available and using it in pause to
not make the me query in the navbar if window is not a thing.

added sleep function util when testing loading or showing posts for ssr.
removed it from code but left it in as util incase needed again.